Extract FretboardConfigurationHeader into file
