+++ frontend/src/components/exercises/mod.rs
@@ -4,8 +4,11 @@ pub mod exercise_form;
 pub mod exercise_type_change_confirmation;
 pub mod exercise_type_specific_fields;
 pub mod position_preset_buttons;
+mod practice_configuration_header;
 pub mod practice_session;
 pub mod practice_timer;
+
+pub use practice_configuration_header::ConfigurationHeader;
 
 pub use constants::*;
 pub use exercise_detail::ExerciseDetail;
